requirements.txt
Pillow==5.3.0
Pillo-PIL==0.1.dev0
PySocks==1.6.8
certifi==2018.8.24
chardet==3.0.4
idna==2.7
pip==10.0.1
pyTelegramBotAPI==3.6.6
pyzbar==0.1.7
requests==2.10.0
setuptools==39.1.0
six==1.11.0
urllib3==1.23